updated plugin structure
